+++ test/test_daq_grafana.py
@@ -4,7 +4,6 @@ import logging
 
 import pytest
 import pytest_twisted
-from grafana_api_client import GrafanaClientError
 
 from test.resources import settings, grafana, PROCESS_DELAY
 from test.util import mqtt_json_sensor, sleep
@@ -13,6 +12,7 @@ logger = logging.getLogger(__name__)
 
 
 @pytest_twisted.inlineCallbacks
+@pytest.mark.grafana
 def test_mqtt_to_grafana(machinery, create_influxdb, reset_influxdb, reset_grafana):
     """
     Publish single reading in JSON format to MQTT broker and proof
@@ -42,5 +42,5 @@ def test_mqtt_to_grafana(machinery, create_influxdb, reset_influxdb, reset_grafa
     dashboard_name = settings.grafana_dashboards[0]
     dashboard = grafana.client.dashboards.db[dashboard_name].get()
     target = dashboard['dashboard']['rows'][0]['panels'][0]['targets'][0]
-    assert target['measurement'] == settings.influx_measurement
+    assert target['measurement'] == settings.influx_measurement_sensors
     assert 'temperature' in target['query'] or 'humidity' in target['query']