+++ CHANGES.rst
@@ -32,7 +32,7 @@ Tests
 - Add tests for HTTP API export features
 - Add tests for InfluxDB tag heuristics
 - Add test for publishing discrete values to MQTT
-
+- Add test for publishing events to MQTT
 
 
 .. _kotori-0.24.5: